packages/simulation/pyproject.toml
[project]
name = "robot-simulation"
version = "0.1.0"
description = "Real-time robot simulation and policy validation package"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "box2d>=2.3.10",
    "pygame>=2.6.1",
    "numpy>=1.24.4",
    "opencv-python>=4.5.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "plyfile>=1.0.0",  # for loading PLY environment files
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.12.3",
    "jupyter-client>=8.6.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
